# Remove inline allocs from smooth and xfrc_accumulate.