--- a/mujoco/mjx/_src/forward.py
+++ b/mujoco/mjx/_src/forward.py
@@ -635,27 +635,18 @@ def fwd_actuation(m: Model, d: Data):
 def fwd_acceleration(m: Model, d: Data):
   """Add up all non-constraint forces, compute qacc_smooth."""
 
-  qfrc_applied = d.qfrc_applied
-  qfrc_accumulated = xfrc_accumulate(m, d)
-
   @kernel
-  def _qfrc_smooth(
-    d: Data,
-    qfrc_applied: wp.array(ndim=2, dtype=wp.float32),
-    qfrc_accumulated: wp.array(ndim=2, dtype=wp.float32),
-  ):
+  def _qfrc_smooth(d: Data):
     worldid, dofid = wp.tid()
     d.qfrc_smooth[worldid, dofid] = (
       d.qfrc_passive[worldid, dofid]
       - d.qfrc_bias[worldid, dofid]
       + d.qfrc_actuator[worldid, dofid]
-      + qfrc_applied[worldid, dofid]
-      + qfrc_accumulated[worldid, dofid]
+      + d.qfrc_applied[worldid, dofid]
     )
 
-  wp.launch(
-    _qfrc_smooth, dim=(d.nworld, m.nv), inputs=[d, qfrc_applied, qfrc_accumulated]
-  )
+  wp.launch(_qfrc_smooth, dim=(d.nworld, m.nv), inputs=[d])
+  xfrc_accumulate(m, d, d.qfrc_smooth)
 
   smooth.solve_m(m, d, d.qacc_smooth, d.qfrc_smooth)