--- a/mujoco/mjx/_src/types.py
+++ b/mujoco/mjx/_src/types.py
@@ -419,6 +419,10 @@ class Data:
   xfrc_applied: wp.array(dtype=wp.spatial_vector, ndim=2)
   contact: Contact
 
+  # arrays used for smooth.rne
+  rne_cacc: wp.array(dtype=wp.spatial_vector, ndim=2)
+  rne_cfrc: wp.array(dtype=wp.spatial_vector, ndim=2)
+
   # temp arrays
   qfrc_integration: wp.array(dtype=wp.float32, ndim=2)
   qacc_integration: wp.array(dtype=wp.float32, ndim=2)