In order to work with JAX graph capture, we must avoid any allocs inside of simulation kernels.

I had to modify `xfrc_accumulate` a bit to remove the mask allocation, but seems to have had no (or perhaps slightly positive) impact on perf.

Result of `mjx-testspeed --function=fwd_acceleration --is_sparse=False --mjcf=humanoid/humanoid.xml --batch_size=8192 --solver=newton --event_trace=True`

This PR:

```
Summary for 8192 parallel rollouts

 Total JIT time: 0.02 s
 Total simulation time: 0.07 s
 Total steps per second: 118,669,204
 Total realtime factor: 593,346.02 x
 Total time per step: 8.43 ns

Event trace:

fwd_acceleration: 6.98
  xfrc_accumulate: 2.63
  solve_m: 3.49
```

Main:

```
Summary for 8192 parallel rollouts

 Total JIT time: 0.56 s
 Total simulation time: 0.07 s
 Total steps per second: 110,029,050
 Total realtime factor: 550,145.25 x
 Total time per step: 9.09 ns

Event trace:

fwd_acceleration: 7.61
  xfrc_accumulate: 3.23
  solve_m: 3.51
```